# app/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "Code With Prince",
      "version": "4.11.1",
      "title": "Code With Prince velaiapp"
     
    },
...
      "servers": [
        {
        "url": "/"
        },{
            "url":"/http"
        }
    ],
...
    "components": {
        "schemas": {
            "TodoList": {
                "type": "object",
                "properties": {
                    "number": {
                    "type": "string"
                    }
                   
                }
            },
...
        }
    },
...
    "paths": {
...
        "/api/seeker_userinfo_details": {
            "post": {
            "tags": [
                "post  API"
            ],
            "summary": "job_seeker_info",
            "requestBody": {
            "description": "Post New todo To Database",
            "required": true,
            "content": {
                
                "application/json": {
                "schema": {
                    "$ref": "#/components61/schemas/TodoList"
                    }
                }
            }
            },
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                    "$ref": "#/components/schemas/TodoList"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/TodoList"
                    }
                }
            }}                        
        },
...
                "/api/longtime_apply_job": {
                    "post": {
                    "tags": [
                        "post  API"
                    ],
                    "summary": "apply_longtime_job",
                    "requestBody": {
                    "description": "Post New todo To Database",
                    "required": true,
                    "content": {
                        
                        "application/json": {
                        "schema": {
                            "$ref": "#/components7/schemas/TodoList"
                            }
                        }
                    }
                    },
                    "responses": {
                        "200": {
                            "description": "OK",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        },
                        "404": {
                            "description": "NOT FOUND",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        }
                    }

                }
                
                
            },
        "/api/shorttime_apply_job": {
            "post": {
            "tags": [
                "post  API"
            ],
            "summary": "apply_longtime_job",
            "requestBody": {
            "description": "Post New todo To Database",
            "required": true,
            "content": {
                
                "application/json": {
                "schema": {
                    "$ref": "#/components8/schemas/TodoList"
                    }
                }
            }
            },
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                    "$ref": "#/components/schemas/TodoList"
                    }
                },
                "404": {
                    "description": "NOT FOUND",
                    "schema": {
                    "$ref": "#/components/schemas/TodoList"
                    }
                }
            }

            }


            },
            "/api/s_like_details": {
                "post": {
                "tags": [
                    "post  API"
                ],
                "summary": "apply_longtime_job",
                "requestBody": {
                "description": "Post New todo To Database",
                "required": true,
                "content": {
                    
                    "application/json": {
                    "schema": {
                        "$ref": "#/components9/schemas/TodoList"
                        }
                    }
                }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                        "$ref": "#/components/schemas/TodoList"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                        "$ref": "#/components/schemas/TodoList"
                        }
                    }
                }
    
                }
    
    
                },
                "/api/l_like_job": {
                    "post": {
                    "tags": [
                        "post  API"
                    ],
                    "summary": "apply_longtime_job",
                    "requestBody": {
                    "description": "Post New todo To Database",
                    "required": true,
                    "content": {
                        
                        "application/json": {
                        "schema": {
                            "$ref": "#/components10/schemas/TodoList"
                            }
                        }
                    }
                    },
                    "responses": {
                        "200": {
                            "description": "OK",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        },
                        "404": {
                            "description": "NOT FOUND",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        }
                    }
        
                }
        
        
            },
                "/api/limit/s_like_apply_check/{user_id}": {
                        "post": {
                        "tags": [
                            "post  API"
                        ],
                        "summary": "apply_longtime_job",
                        "parameters": [{
                            "name": "user_id",
                            "in": "path",
                            "description": "job_apply_count",
                            "required": true,
                            "type": "integer",
                            "schema" : {
                                "$ref": "#/components/schemas/Id"
                            }
                            }],
                        "requestBody": {
                        "description": "Post New todo To Database",
                        "required": true,
                        "content": {
                            
                            "application/json": {
                            "schema": {
                                "$ref": "#/components11/schemas/TodoList"
                                }
                            }
                        }
                        },
                        "responses": {
                            "200": {
                                "description": "OK",
                                "schema": {
                                "$ref": "#/components/schemas/TodoList"
                                }
                            },
                            "404": {
                                "description": "NOT FOUND",
                                "schema": {
                                "$ref": "#/components/schemas/TodoList"
                                }
                            }
                        }
            
                    }
                },
                        "/api/limit/L_like_apply_check/{user_id}": {
                            "post": {
                            "tags": [
                                "post  API"
                            ],
                            "summary": "user_apply_like_check_show",
                            "parameters": [{
                                "name": "user_id",
                                "in": "path",
                                "description": "job_apply_count",
                                "required": true,
                                "type": "integer"
                                
                                }],
                            "requestBody": {
                            "description": "Post New todo To Database",
                            "required": true,
                            "content": {
                                
                                "application/json": {
                                "schema": {
                                    "$ref": "#/components11/schemas/TodoList"
                                    }
                                }
                            }
                            },
                            "responses": {
                                "200": {
                                    "description": "OK",
                                    "schema": {
                                    "$ref": "#/components/schemas/TodoList"
                                    }
                                },
                                "404": {
                                    "description": "NOT FOUND",
                                    "schema": {
                                    "$ref": "#/components/schemas/TodoList"
                                    }
                                }
                            }
                
                            }
                
                
                            },
...
                "/api/l_like_job/{user_id}":{

                    "get" :{
                        "tags": [
                            "Get  API"
                        ],
                        "summary": "Get ",
                        "parameters": [{
                        "name": "user_id",
                        "in": "path",
                        "description": "job_apply_count",
                        "required": true,
                        "type": "integer",
                        "schema" : {
                            "$ref": "#/components/schemas/Id"
                        }
                        }],
                    
                        "responses": {
                        "200": {
                            "description": "OK",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        },
                        "404": {
                            "description": "NOT FOUND",
                            "schema": {
                            "$ref": "#/components/schemas/TodoList"
                            }
                        }
                    }
                }
            },
...
    }
}